Add files via upload
Assignment 23: Write a program to initialize an integer array and print the sum and average of the array.
Assignment 24: Write a program to initialize an integer array and find the maximum and minimum value of the array.
Assignment 25: Write a program to initialize an integer array with values and check if a given number is present in the array or not.
Assignment 26: Initialize an integer array with ASCII values and print the corresponding character values in a single row.
Assignment 27: Write a program to find the largest 2 numbers and the smallest 2 numbers in the given array.
Assignment 28: Write a program to initialize an array and print them in sorted order.
Assignment 29: Write a program to remove the duplicate elements in an array and print the same.
Assignment 30: Write a program to print the sum of the elements of an array. If the array has 6 and 7 in succeeding orders, ignore the numbers between 6 and 7 and consider the other numbers for calculation of sum.
Assignment 31: Write a program to reverse the elements of a given 2*2 array. Four integer numbers need to be passed as Command Line Arguments.
Assignment 32: Write a program to find the biggest number in a 3*3 array. The program is supposed to receive 9 integer numbers as command line arguments.